[Rename] vislib plugin
