--- src/plugins/vis_type_vislib/kibana.json
@@ -1,9 +1,0 @@
-{
-  "id": "visTypeVislib",
-  "version": "kibana",
-  "server": true,
-  "ui": true,
-  "requiredPlugins": ["charts", "data", "expressions", "visualizations", "kibanaLegacy"],
-  "optionalPlugins": ["visTypeXy"],
-  "requiredBundles": ["kibanaUtils", "visDefaultEditor"]
-}
