+++ src/plugins/vis_type_vislib/opensearch_dashboards.json
@@ -1,0 +1,9 @@
+{
+  "id": "visTypeVislib",
+  "version": "opensearchDashboards",
+  "server": true,
+  "ui": true,
+  "requiredPlugins": ["charts", "data", "expressions", "visualizations", "opensearchDashboardsLegacy"],
+  "optionalPlugins": ["visTypeXy"],
+  "requiredBundles": ["opensearchDashboardsUtils", "visDefaultEditor"]
+}
