+++ src/plugins/vis_type_vislib/public/components/common/truncate_labels.tsx
@@ -18,7 +18,7 @@
  */
 
 import React, { ChangeEvent } from 'react';
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 import { EuiFormRow, EuiFieldNumber } from '@elastic/eui';
 
 interface TruncateLabelsOptionProps {
