+++ src/plugins/vis_type_vislib/public/components/options/gauge/ranges_panel.tsx
@@ -19,8 +19,8 @@
 
 import React, { useCallback } from 'react';
 import { EuiPanel, EuiSpacer, EuiTitle } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 
 import {
   ColorRanges,
