+++ src/plugins/vis_type_vislib/public/components/options/metrics_axes/category_axis_panel.tsx
@@ -20,8 +20,8 @@
 import React, { useCallback } from 'react';
 
 import { EuiPanel, EuiTitle, EuiSpacer } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 
 import { VisOptionsProps } from 'src/plugins/vis_default_editor/public';
 import { Axis } from '../../../types';
