+++ src/plugins/vis_type_vislib/public/components/options/metrics_axes/custom_extents_options.tsx
@@ -18,7 +18,7 @@
  */
 
 import React, { useCallback, useEffect } from 'react';
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 
 import { ValueAxis } from '../../../types';
 import { NumberInputOption, SwitchOption } from '../../../../../charts/public';
