+++ src/plugins/vis_type_vislib/public/components/options/metrics_axes/series_panel.tsx
@@ -20,8 +20,8 @@
 import React from 'react';
 
 import { EuiPanel, EuiTitle, EuiSpacer, EuiAccordion } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 
 import { Vis } from '../../../../../visualizations/public';
 import { ValueAxis, SeriesParam } from '../../../types';
