+++ src/plugins/vis_type_vislib/public/components/options/point_series/grid_panel.tsx
@@ -19,8 +19,8 @@
 import React, { useMemo, useEffect, useCallback } from 'react';
 
 import { EuiPanel, EuiTitle, EuiSpacer } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 
 import { VisOptionsProps } from 'src/plugins/vis_default_editor/public';
 import { SelectOption, SwitchOption } from '../../../../../charts/public';
