+++ src/plugins/vis_type_vislib/public/components/options/point_series/threshold_panel.tsx
@@ -18,8 +18,8 @@
  */
 import React, { useCallback } from 'react';
 import { EuiPanel, EuiTitle, EuiColorPicker, EuiFormRow, EuiSpacer } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 
 import { ValidationVisOptionsProps } from '../../common';
 import {
