+++ src/plugins/vis_type_vislib/public/histogram.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 // @ts-ignore
 import { palettes } from '@elastic/eui/lib/services';
 // @ts-ignore
