+++ src/plugins/vis_type_vislib/public/horizontal_bar.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 // @ts-ignore
 import { palettes, euiPaletteColorBlind } from '@elastic/eui/lib/services';
