+++ src/plugins/vis_type_vislib/public/utils/collections.ts
@@ -17,8 +17,8 @@
  * under the License.
  */
 
-import { i18n } from '@kbn/i18n';
-import { $Values } from '@kbn/utility-types';
+import { i18n } from '@osd/i18n';
+import { $Values } from '@osd/utility-types';
 
 import { colorSchemas, Rotates } from '../../../charts/public';
