+++ src/plugins/vis_type_vislib/public/utils/common_config.tsx
@@ -18,7 +18,7 @@
  */
 
 import React from 'react';
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 
 import { VisOptionsProps } from 'src/plugins/vis_default_editor/public';
 import { PointSeriesOptions, MetricsAxisOptions } from '../components/options';
