+++ src/plugins/vis_type_vislib/public/vislib/components/labels/data_array.js
@@ -21,7 +21,7 @@ import _ from 'lodash';
 import { flattenSeries } from './flatten_series';
 
 /*
- * Accepts a Kibana data object and returns an array of values objects.
+ * Accepts a OpenSearch Dashboards data object and returns an array of values objects.
  */
 export function dataArray(obj) {
   if (!_.isObject(obj) || (!obj.rows && !obj.columns && !obj.series)) {
