+++ src/plugins/vis_type_vislib/public/vislib/components/legend/legend_item.tsx
@@ -20,8 +20,8 @@
 import React, { memo, BaseSyntheticEvent, KeyboardEvent } from 'react';
 import classNames from 'classnames';
 
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n/react';
+import { i18n } from '@osd/i18n';
+import { FormattedMessage } from '@osd/i18n/react';
 import {
   EuiPopover,
   keys,
