+++ src/plugins/vis_type_vislib/public/vislib/components/tooltip/_hierarchical_tooltip_formatter.js
@@ -22,7 +22,7 @@ import _ from 'lodash';
 import numeral from '@elastic/numeral';
 import { renderToStaticMarkup } from 'react-dom/server';
 
-import { FormattedMessage, I18nProvider } from '@kbn/i18n/react';
+import { FormattedMessage, I18nProvider } from '@osd/i18n/react';
 
 import { collectBranch } from './_collect_branch';
