+++ src/plugins/vis_type_vislib/public/vislib/components/zero_injection/flatten_data.js
@@ -20,7 +20,7 @@
 import _ from 'lodash';
 
 /*
- * Accepts a Kibana data object, flattens the data.series values array,
+ * Accepts a OpenSearch Dashboards data object, flattens the data.series values array,
  * and returns an array of values objects.
  */
 export function flattenData(obj) {
