+++ src/plugins/vis_type_vislib/public/vislib/helpers/point_series/_fake_x_aspect.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 import { Aspect } from './point_series';
 
 export function makeFakeXAspect() {
