+++ src/plugins/vis_type_vislib/public/vislib/lib/handler.js
@@ -29,7 +29,7 @@ import { Axis } from './axis/axis';
 import { ChartGrid as Grid } from './chart_grid';
 import { visTypes as chartTypes } from '../visualizations/vis_types';
 import { Binder } from './binder';
-import { dispatchRenderComplete } from '../../../../kibana_utils/public';
+import { dispatchRenderComplete } from '../../../../opensearch_dashboards_utils/public';
 
 const markdownIt = new MarkdownIt({
   html: false,
