+++ src/plugins/vis_type_vislib/public/vislib/lib/types/point_series.js
@@ -18,7 +18,7 @@
  */
 
 import _ from 'lodash';
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 
 function getSeriId(seri) {
   if (!seri.id) {
